+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.5.0] - 2024-05-02
 
+### Fixed
+
+- False positive `FormatStringValid` issues on strings containing uppercase format specifiers.
+
 ### Added
 
 - Support for the `winapi` calling convention.